feat: use latest api

+++ bungeecord/src/main/java/com/vexsoftware/votifier/bungee/BungeeScheduler.java
@@ -14,12 +14,12 @@ class BungeeScheduler implements VotifierScheduler {
     }
 
     @Override
-    public ScheduledVotifierTask delayedOnPool(Runnable runnable, int delay, TimeUnit unit) {
+    public ScheduledVotifierTask delayedOnPool(Runnable runnable, long delay, TimeUnit unit) {
         return new BungeeTaskWrapper(plugin.getProxy().getScheduler().schedule(plugin, runnable, delay, unit));
     }
 
     @Override
-    public ScheduledVotifierTask repeatOnPool(Runnable runnable, int delay, int repeat, TimeUnit unit) {
+    public ScheduledVotifierTask repeatOnPool(Runnable runnable, long delay, long repeat, TimeUnit unit) {
         return new BungeeTaskWrapper(plugin.getProxy().getScheduler().schedule(plugin, runnable, delay, repeat, unit));
     }
